refactor(client/atoms): removed textinput atom and transfered logic to login component
