+++ README.md
@@ -11,3 +11,5 @@ To use expo:
 Links:
 Trello:
 https://trello.com/b/Zguti6wp/tagdate%F0%9F%8F%84%E2%99%80%EF%B8%8F%F0%9F%8F%84%E2%99%82%EF%B8%8F-thesis-project
+
+testing commitizen...
