+++ client/components/organisms/Login.tsx
@@ -1,25 +1,28 @@
 import React, { useState } from "react";
 import { View, Text, StyleSheet } from "react-native";
+import { TextInput } from "react-native";
 import AppButton from "../atoms/AppButton";
-import AppTextInput from "../atoms/AppTextInput";
+import colors from "../../helpers/colors";
 
 function Login(props) {
-  console.log(`props`, props);
-
   const [username, setUsername] = useState("");
   const [password, setPassword] = useState("");
 
   console.log("username", username, "password", password);
 
   const handleLogin = () => {
-    // console.log("pressing login button");
     props.navigation.replace("LoginSuccessPage");
   };
 
   return (
     <View style={styles.loginContainer}>
-      <AppTextInput placeholder="username" handler={setUsername} />
-      <AppTextInput
+      <TextInput
+        style={styles.textInput}
+        placeholder="username"
+        onChangeText={setUsername}
+      />
+      <TextInput
+        style={styles.textInput}
         placeholder="password"
         onChangeText={setPassword}
         secureTextEntry={true}
@@ -34,6 +37,14 @@ const styles = StyleSheet.create({
     alignItems: "center",
     justifyContent: "center",
     width: "100%",
+  },
+  textInput: {
+    margin: 10,
+    textAlign: "center",
+    borderBottomColor: colors.violet,
+    borderBottomWidth: 2,
+    width: "80%",
+    color: colors.violet,
   },
 });
